BoundaryConditions

ParallelTemperingMonteCarlo.BoundaryConditions — Module

module BoundaryConditions

This module provides structs and methods for different kinds of boundary conditions.

ParallelTemperingMonteCarlo.BoundaryConditions.PositionArray — Type

PositionArray = Union{Vector{Vector{T}}, Vector{SVector{3, T}}} where T <: Number

Type alias for a list of positions.

ParallelTemperingMonteCarlo.BoundaryConditions.PositionVector — Type

PositionVector = Union{SVector{3, T}, Vector{T}} where T <: Number

Type alias for all kinds of acceptable position vectors.

ParallelTemperingMonteCarlo.BoundaryConditions.AbstractBC — Type

AbstractBC{T}

Encompasses possible boundary conditions; implemented:

SphericalBC

PeriodicBC

Needs methods implemented for

atom_displacement

ParallelTemperingMonteCarlo.BoundaryConditions.CubicBC — Type

CubicBC{T}(; side_length::Number)

Subtype of periodic boundary conditions where the box_length is isotropic.

ParallelTemperingMonteCarlo.BoundaryConditions.PeriodicBC — Type

PeriodicBC{T}

Overarching type of boundary condition for simulating the infinite bulk

Implemented types:

CubicBC

RhombicBC

ParallelTemperingMonteCarlo.BoundaryConditions.RhombicBC — Type

RhombicBC{T}(; length::Number, height::Number)

Subtype of periodic boundary condition where the box_length and box_height are not the same. The projection of the box on the xy-plane is a rhombus, box_length applies to all four sides.

ParallelTemperingMonteCarlo.BoundaryConditions.SphericalBC — Type

SphericalBC{T}(;radius::Number)

Implements type for spherical boundary conditions; subtype of AbstractBC. Needs radius of binding sphere as keyword argument. Fieldname: radius2: squared radius of binding sphere

ParallelTemperingMonteCarlo.BoundaryConditions.check_boundary — Method

check_boundary(bc::SpericalBC,pos::PositionVector) where T <: Number

Returns true when atom outside of spherical boundary (squared norm of position vector < radius^2 of binding sphere).

ParallelTemperingMonteCarlo.BoundaryConditions.test_cluster_inside — Method

test_cluster_inside(pos::Vector{SVector{3,T}},bc::SphericalBC) where T <: Number

Tests if whole cluster lies in the binding sphere.