CustomTypes

ParallelTemperingMonteCarlo.CustomTypes.PositionArray — Type

PositionArray = Union{Vector{Vector{T}}, Vector{SVector{3, T}}} where T <: Number

Type alias for a list of positions.

ParallelTemperingMonteCarlo.CustomTypes.PositionVector — Type

PositionVector = Union{SVector{3, T}, Vector{T}} where T <: Number

Type alias for all kinds of acceptable position vectors.

ParallelTemperingMonteCarlo.CustomTypes.VorS — Type

VorS = T where T <: AbstractArray{Z, 1} where Z <: Number

Type alias for a collection of numbers. The name derives from most collections being a Vector or StaticVector of numbers.