Initialization

ParallelTemperingMonteCarlo.Initialization.initialisation — Method

initialisation(mc_params::MCParams,temp::TempGrid,start_config::Config,potential::Ptype,ensemble::NVT)
initialisation(restart::Bool;eq_cycles::Number)

Basic function for establishing the structs and parameters required for the simulation. Inputs for method one are:

mc_params: the basic values and parameters concerning how long our simulation runs.

temp: a grid of temp and beta values passed to the mc_states struct.

start_config: the initial configuration used to populate each starting state with.

potential: the potential energy used for the simulation.

ensemble: the ensemble used for the simulation, contains the move strat inherently.

Inputs for method two are:

restart: A boolean determining whether the simulation is restarting or being read-in from a file.

eq_cycles: the proportion of n_cycles to be run in equilibration.

Method one and two return the following structs:

mc_states: a vector of MCState structs each at a different temperature

move_strategy: struct containing a vector of MoveType

results: struct countaining the output such as Cv and histograms

start_counter: where to begin the sims

n_steps: total moves per mc_cycle

Method two also returns:

mc_params: the static parameters determining the scope of the simulation

ensemble: determines the ensemble and move_strategy followed by the simulation

potential: the potential energy surface explored by trajectories

***NOTES FOR FUTURE IMPLEMENTATION***

consider shuffling mc_params to include the tempgrid and cut down the number of inputs.