InputParams

ParallelTemperingMonteCarlo.InputParams — Module

module InputParams

This module provides structs and methods to arrange input parameters.

ParallelTemperingMonteCarlo.InputParams.MCParams — Type

MCParams(cycles::Int, n_traj::Int, n_atoms::Int; eq_percentage = 0.2, mc_sample = 1, n_adjust = 100, n_bin = 100)

Type that collects MC specific data; field names:

mc_cycle::Int: number of MC cycles

eq_cycles::Int: number of equilibration cycles (default 20% of mc_cycle)

mc_sample::Int: gives number of MC cycles after which energy is saved (default: 1)

n_traj::Int: number of trajectories (ie. temperatures) propagated in parallel

n_atoms::Int: number N of atoms in configuration

n_adjust::Int: number of moves after which step size of atom/volume moves is adjusted (default: 100)

n_bin::Int: number of histogram bins (default: 100)

ParallelTemperingMonteCarlo.InputParams.Output — Type

Output{T}(n_bin; en_min = 0) where T <: Number

Collects output of MC calculation; field names:

n_bin::Int: number of energy bins for histograms

en_min::T: minimum energy found during calculation

en_max::T: maximum energy found during calculation

v_min::T: minimum volume

v_max::T: maximum volume

delta_en_hist::T: the step size associated with the energy histogram

delta_v_hist::T: step size associated with volume histogram

delta_r2::T: step size associated with the rdf histogram

max_displ::Vector{T}: final maximum displacements

en_avg::Vector{T}: inner energy U(T) (as average over sampled energies)

heat_cap::Vector{T}: heat capacities C(T)

rdf::Vector{Vector{T}}: radial distribution information

count_stat_*::Vector{T}: statistics of accepted atom, volume and rotation moves and attempted and successful parallel-tempering exchanges

ParallelTemperingMonteCarlo.InputParams.TempGrid — Type

TempGrid{N}(ti::Number, tf::Number; tdistr) 
TempGrid(ti::Number, tf::Number, N::Int; tdistr=:geometric)

Generates grid of N temperatures and inverse temperatures for MC calculation between initial and final temperatures ti and tf.

Field names:

t_grid::SVector{N,T}: temperatures (in K)

beta_grid::SVector{N,T}: inverse temperatures (in atomic units)

Keyword argument tdistr:

:geometric (default): generates geometric temperature distribution

:equally_spaced: generates equally spaced temperature grid (not implemented presently)