MCMoves

ParallelTemperingMonteCarlo.MCMoves.atom_displacement — Method

atom_displacement(pos::PositionVector, max_displacement::Number, bc::SphericalBC)
atom_displacement(pos::PositionVector, max_displacement::Number, bc::CubicBC)
atom_displacement(pos::PositionVector, max_displacement::Number, bc::RhombicBC)
atom_displacement(mc_state::MCState{T, N, BC}) where {T, N, BC <: PeriodicBC}
atom_displacement(mc_state::MCState{T, N, BC}) where {T, N, BC <: SphericalBC}

Generates trial position for atom, moving it from pos by some random displacement. Random displacement determined by max_displacement. These variables are additionally contained in mc_state where the pos is determined by index. Implemented for:

SphericalBC: trial move is repeated until moved atom is within binding sphere

CubicBC; RhombicBC; RectangularBC: periodic boundary condition enforced, an atom is moved into the box from the other side when it tries to get out.

The final method is a wrapper function which unpacks mc_states, which contains all the necessary arguments for the two methods above. When we have correctly implemented move_strat this wrapper will be expanded to include other methods.

ParallelTemperingMonteCarlo.MCMoves.generate_move! — Method

generate_move!(mc_state::MCState,movetype::String)

generate_move! is the currying function that takes mc_state and a movetype and generates the variables required inside of the ensemblevariables struct within mc_state.

ParallelTemperingMonteCarlo.MCMoves.scale_xy — Method

scale_xy(::RhombicBC, α)
scale_xy(::RectangularBC, α)
scale_xy(::Vector{<:SVector}, α)
scale_xy(::Config, α)

Scale boundary condition, vector, or configuration in all $x$ and $y$ dimensions by factor α.

ParallelTemperingMonteCarlo.MCMoves.scale_xyz — Method

scale_xyz(::RhombicBC, α)
scale_xyz(::RectangularBC, α)
scale_xyz(::Vector{<:SVector}, α)
scale_xyz(::Config, α)

Scale boundary condition, vector, or configuration in all three dimensions by factor α.

ParallelTemperingMonteCarlo.MCMoves.scale_z — Method

scale_z(::RhombicBC, α)
scale_z(::RectangularBC, α)
scale_z(::Vector{<:SVector}, α)
scale_z(::Config, α)

Scale boundary condition, vector, or configuration in the $z$ dimension by factor α.

ParallelTemperingMonteCarlo.MCMoves.swap_atoms — Method

(swap_atoms(mc_state::MCState{T, N, BC, PV, EV}) where {T, N, BC, PV, EV <: NNVTVariables{tee, n, N1, N2}}) where {tee, n, N1, N2}

Swaps two atoms in the configuration.

ParallelTemperingMonteCarlo.MCMoves.volume_change — Function

volume_change(mc_state::MCState, separated_volume=false)

MC move that changes volume. If separated_volume == true, the volume is changed in the $x$,$y$ directions or in the $z$ direction separately.

ParallelTemperingMonteCarlo.MCMoves.volume_change_separated — Method

volume_change_separated(mc_state::MCState)

Change the volume

in the $z$-direction with probability $1/3$,

in the $x$,$y$-directions with probability $2/3$ or

unioformly with probability $1/2$.

Update the mc_state accordingly. If the potential used is ELJPotentialB or LookupTablePotential, update the tangent matrix as well.

ParallelTemperingMonteCarlo.MCMoves.volume_change_uniform — Method

volume_change_uniform(mc_state::MCState)

Change the volume uniformly and update the mc_state accordingly.

ParallelTemperingMonteCarlo.MCMoves.volume_change_xy — Method

volume_change_xy(conf::Config, bc, max_vchange::Real, max_length::Real, lh_ratio)

Scale the whole configuration, including positions and the box length by a random amount in the $x$ and $y$ directions.

Returns the trial configuration.

ParallelTemperingMonteCarlo.MCMoves.volume_change_xyz — Method

volume_change_xyz(conf::Config, bc, max_vchange::Real, max_length::Real)

Scale the whole configuration, including positions and the box length by a random amount. Returns the trial configuration.

ParallelTemperingMonteCarlo.MCMoves.volume_change_z — Method

volume_change_z(conf::Config, max_vchange::Real, max_length::Real, lh_ratio)

Scale the whole configuration, including positions and the box length by a random amount in the $z$ direction.

Returns the trial configuration.